osm/requirements.txt
neo4j==5.7.0
psycopg==3.1.8
python-dotenv==1.0.0
